# models/state.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// State state
//
// swagger:model State
type State struct {

	// i s o code
	ISOCode string `json:"iso_code,omitempty"`

	// name
	Name string `json:"name,omitempty"`
}

// Validate validates this state
func (m *State) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this state based on context it is used
func (m *State) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *State) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *State) UnmarshalBinary(b []byte) error {
	var res State
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}