# swagger.json
{
  "swagger": "2.0",
  "info": {},
  "paths": {
...
    "/app/express-checkout": {
      "post": {
        "description": "This endpoint is for express checkout.\n\nEndpoint: /app/express-checkout\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "ExpressCheckout"
        ],
        "summary": "expressCheckout",
        "operationId": "expressCheckout",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ExpressCheckoutOpts"
            }
          },
          {
            "type": "string",
            "description": "Platform type for example android, web or ios.",
            "name": "platform",
            "in": "query",
            "required": true
          },
          {
            "description": "Customer login required for successful response.",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "description": "Token required for successful response.",
            "name": "auth token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " OK",
            "schema": {
              "$ref": "#/definitions/OrderInfo"
            }
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": " Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
    "/app/express-checkout/check/cod": {
      "post": {
        "description": "This endpoint will express checkout RTO.\n\nEndpoint: /app/express-checkout/check/cod\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "CODviaGoKwik"
        ],
        "summary": "expressCheckoutRTO",
        "operationId": "expressCheckoutRTO",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ExpressCheckoutWebOpts"
            }
          },
          {
            "description": "Customer login required for successful response.",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "description": "Token required for successful response.",
            "name": "auth token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " OK"
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
...
    "/app/wishlist": {
      "put": {
        "description": "This endpoint will add the product to the wishlist.\n\nEndpoint: /app/wishlist\n\nMethod: PUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Wishlist"
        ],
        "summary": "addToWishlist",
        "operationId": "addToWishlist",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/AddToWishlistOpts"
            }
          },
          {
            "description": "Customer login required for successful response.",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "description": "Token required for successful response.",
            "name": "auth token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " OK"
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": " Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      },
      "delete": {
        "description": "This endpoint will delete the product from the wishlist.\n\nEndpoint: /app/wishlist\n\nMethod: DELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Wishlist"
        ],
        "summary": "removeFromWishlist",
        "operationId": "removeFromWishlist",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/RemoveFromWishlistOpts"
            }
          },
          {
            "description": "Customer login required for successful response.",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "description": "Token required for successful response.",
            "name": "auth token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " true"
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": " Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
    "/app/wishlist/{userID}": {
      "get": {
        "description": "This endpoint will return the wishlist product of the user.\n\nEndpoint: /app/wishlist/{userID}\n\nMethod: GET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Wishlist"
        ],
        "summary": "getWishlist",
        "operationId": "getWishlist",
        "parameters": [
          {
            "enum": [
              "6065d4503824bf77961c21ae"
            ],
            "type": "object",
            "name": "userID",
            "in": "path",
            "required": true
          },
          {
            "description": "Customer login required for successful response.",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "description": "Token required for successful response.",
            "name": "auth token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " true"
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": " Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
...
    "/web/express-checkout": {
      "post": {
        "description": "This endpoint is for express checkout for web.\n\nEndpoint: /web/express-checkout\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "ExpressCheckout"
        ],
        "summary": "expressCheckoutWeb",
        "operationId": "expressCheckoutWeb",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ExpressCheckoutWebOpts"
            }
          },
          {
            "type": "string",
            "description": "Platform type for example android, web or ios.",
            "name": "platform",
            "in": "query",
            "required": true
          },
          {
            "description": "Customer login required for successful response.",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "description": "Token required for successful response.",
            "name": "auth token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " OK",
            "schema": {
              "$ref": "#/definitions/OrderInfo"
            }
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": " Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "AddToWishlistOpts": {
      "type": "object",
      "properties": {
        "catalog_id": {
          "$ref": "#/definitions/ObjectID"
        },
        "user_id": {
          "$ref": "#/definitions/ObjectID"
        }
      },
      "x-go-package": "go-app/schema"
    },
...
    "AppErr": {
      "type": "object",
      "properties": {
        "Error": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "RequestID": {
          "type": "string"
        }
      },
      "x-go-package": "go-app/server/handler"
    },
...
    "ExpressCheckoutOpts": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/OrderAddressOpts"
        },
        "coupon": {
          "type": "string",
          "x-go-name": "Coupon"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ItemExpress"
          },
          "x-go-name": "Items"
        },
        "source": {
          "type": "string",
          "x-go-name": "Source"
        },
        "source_id": {
          "$ref": "#/definitions/ObjectID"
        },
        "user_id": {
          "$ref": "#/definitions/ObjectID"
        }
      },
      "x-go-package": "go-app/schema"
    },
    "ExpressCheckoutWebOpts": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/OrderAddressOpts"
        },
        "coupon": {
          "type": "string",
          "x-go-name": "Coupon"
        },
        "is_cod": {
          "type": "boolean",
          "x-go-name": "IsCOD"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ItemExpress"
          },
          "x-go-name": "Items"
        },
        "request_id": {
          "type": "string",
          "x-go-name": "RequestID"
        },
        "source": {
          "type": "string",
          "x-go-name": "Source"
        },
        "source_id": {
          "$ref": "#/definitions/ObjectID"
        },
        "user_id": {
          "$ref": "#/definitions/ObjectID"
        }
      },
      "x-go-package": "go-app/schema"
    },
...
    "GoKwikCreateOrderResp": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/GoKwikOrderData"
        },
        "statusCode": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "StatusCode"
        },
        "statusMessage": {
          "type": "string",
          "x-go-name": "StatusMessage"
        }
      },
      "x-go-package": "go-app/schema"
    },
    "GoKwikOrderData": {
      "type": "object",
      "properties": {
        "gokwik_oid": {
          "type": "string",
          "x-go-name": "GokwikOID"
        },
        "mid": {
          "type": "string",
          "x-go-name": "Mid"
        },
        "moid": {
          "type": "string",
          "x-go-name": "Moid"
        },
        "order_status": {
          "type": "string",
          "x-go-name": "OrderStatus"
        },
        "order_type": {
          "type": "string",
          "x-go-name": "OrderType"
        },
        "phone": {
          "type": "string",
          "x-go-name": "Phone"
        },
        "request_id": {
          "type": "string",
          "x-go-name": "RequestID"
        },
        "total": {
          "type": "string",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "go-app/schema"
    },
...
    "ItemExpress": {
      "type": "object",
      "properties": {
        "catalog_id": {
          "$ref": "#/definitions/ObjectID"
        },
        "quantity": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Quantity"
        },
        "source": {
          "$ref": "#/definitions/Source"
        },
        "variant_id": {
          "$ref": "#/definitions/ObjectID"
        }
      },
      "x-go-package": "go-app/schema"
    },
...
    "ObjectID": {
      "type": "array",
      "title": "ObjectID is the BSON ObjectID type.",
      "items": {
        "type": "integer",
        "format": "uint8"
      },
      "x-go-package": "go.mongodb.org/mongo-driver/bson/primitive"
    },
    "OrderAddressOpts": {
      "description": "OrderAddressOpts contains field required to add/edit the address of the user's cart",
      "type": "object",
      "properties": {
        "city": {
          "type": "string",
          "x-go-name": "City"
        },
        "display_name": {
          "type": "string",
          "x-go-name": "DisplayName"
        },
        "district": {
          "type": "string",
          "x-go-name": "District"
        },
        "id": {
          "$ref": "#/definitions/ObjectID"
        },
        "line1": {
          "type": "string",
          "x-go-name": "Line1"
        },
        "line2": {
          "type": "string",
          "x-go-name": "Line2"
        },
        "phone_no": {
          "$ref": "#/definitions/PhoneNumber"
        },
        "plain_address": {
          "type": "string",
          "x-go-name": "PlainAddress"
        },
        "postal_code": {
          "type": "string",
          "x-go-name": "PostalCode"
        },
        "state": {
          "$ref": "#/definitions/State"
        }
      },
      "x-go-package": "go-app/schema"
    },
    "OrderInfo": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "number",
          "format": "float",
          "x-go-name": "Amount"
        },
        "go_kwik": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GoKwikCreateOrderResp"
          },
          "x-go-name": "GoKwikResponse"
        },
        "order_id": {
          "type": "string",
          "x-go-name": "OrderID"
        },
        "razorpay_id": {
          "type": "string",
          "x-go-name": "RazorpayID"
        }
      },
      "x-go-package": "go-app/schema"
    },
...
    "PhoneNumber": {
      "description": "PhoneNumber represents a contact number contains prefix (country code) and phone number",
      "type": "object",
      "properties": {
        "number": {
          "description": "number",
          "type": "string",
          "x-go-name": "Number"
        },
        "prefix": {
          "description": "prefix",
          "type": "string",
          "x-go-name": "Prefix"
        }
      },
      "x-go-package": "go-app/model"
    },
...
    "RemoveFromWishlistOpts": {
      "type": "object",
      "properties": {
        "catalog_id": {
          "$ref": "#/definitions/ObjectID"
        },
        "user_id": {
          "$ref": "#/definitions/ObjectID"
        }
      },
      "x-go-package": "go-app/schema"
    },
...
    "Source": {
      "description": "Source source",
      "type": "object",
      "properties": {
        "id": {
          "description": "ID",
          "type": "string",
          "x-go-name": "ID"
        },
        "type": {
          "description": "type",
          "type": "string",
          "x-go-name": "Type"
        }
      },
      "x-go-package": "go-app/models"
    },
...
    "State": {
      "description": "State state",
      "type": "object",
      "properties": {
        "iso_code": {
          "description": "i s o code",
          "type": "string",
          "x-go-name": "ISOCode"
        },
        "name": {
          "description": "name",
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "go-app/models"
    },
...
  }
}